+++ eBudgeting/src/main/java/biz/thaicom/eBudgeting/repositories/AssetAllocationRepository.java
@@ -17,7 +17,7 @@ public interface AssetAllocationRepository extends JpaRepository<AssetAllocation
 			"SELECT sum(assetAllocation.quantity * assetAllocation.unitBudget) " +
 			"FROM AssetAllocation assetAllocation " +
 			"WHERE assetAllocation.proposal = ?1 ")
-	Long findSumBudgetOfPropsoal(BudgetProposal proposal);
+	Double findSumBudgetOfPropsoal(BudgetProposal proposal);
 
 	@Query("" +
 			"SELECT assetAllocation " +
